Moved conundrum logic to server function, allowing inports in server functions
