+++ src/LettersWorker.ts
@@ -1,7 +1,22 @@
-import { findWords } from "./Anagrams";
-
 function postLettersResult(msg: string[]) {
     postMessage(msg);
+}
+
+async function findWords(letters: string, maxResults: number) {
+    const response = await fetch('/.netlify/functions/findWords', {
+        method: 'POST',
+        body: JSON.stringify({
+            letters: letters.toLowerCase(),
+            maxResults
+        })
+    });
+
+    if (!response.ok) {
+        return [];
+    }
+
+    const data: string[] = await response.json();
+    return data;
 }
 
 let bestWords: string[] = ['Got no', 'response'];
